medicine_recognition_app/requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
easyocr==1.6.2
filelock==3.12.0
idna==3.4
imageio==2.28.1
inflection==0.5.1
Jinja2==3.1.2
jsonschema==4.17.3
lazy-loader==0.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.24.3
opencv-python-headless==4.5.4.60
packaging==23.1
Pillow==9.5.0
pyclipper==1.3.0.post4
pyrsistent==0.19.3
python-bidi==0.4.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.30.0
scikit-image==0.20.0
scipy==1.10.1
shapely==2.0.1
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tifffile==2023.4.12
torch==2.0.1
torchvision==0.15.2
typing-extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
